refactor: Problem.QuestionType -> utils.QuestionType
