+++ src/main/java/code/rice/bowl/spaghetti/dto/problem/ProblemDetailResponse.java
@@ -1,5 +1,6 @@
 package code.rice.bowl.spaghetti.dto.problem;
 
+import code.rice.bowl.spaghetti.utils.QuestionType;
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,6 +14,7 @@ import lombok.Setter;
 @Builder
 public class ProblemDetailResponse {
     Long problemId;
+    QuestionType type;
     String title;
     JsonNode description;
 }
